requirements.txt
Pillow==10.0.0
Pillow==10.0.0
streamlit==1.25.0
